`invitationsView`: The query result must be scanned from the response structure. Otherwise, the format is not right (boolean are converted as "0" and "1", instead of true and false).
