+++ app/doc/models.go
@@ -51,44 +51,3 @@ type groupsMembershipHistoryResponseRow struct {
 		Type string `json:"type"`
 	} `json:"group"`
 }
-
-// swagger:model invitationsViewResponseRow
-type invitationsViewResponseRow struct {
-	// `group_membership_changes.group_id`
-	// required: true
-	GroupID int64 `json:"group_id"`
-	// `groups_groups.type_changed_at`
-	// required: true
-	At time.Time `json:"at"`
-
-	// the user that invited
-	// required: true
-	InvitingUser struct {
-		// `users.group_id`
-		// required: true
-		ID int64 `json:"id"`
-		// required: true
-		Login string `json:"login"`
-		// Nullable
-		// required: true
-		FirstName string `json:"first_name"`
-		// Nullable
-		// required: true
-		LastName string `json:"last_name"`
-	} `json:"inviting_user"`
-
-	// required: true
-	Group struct {
-		// `groups.id`
-		// required: true
-		ID int64 `json:"id"`
-		// required: true
-		Name string `json:"name"`
-		// Nullable
-		// required: true
-		Description *string `json:"description"`
-		// required: true
-		// enum: Class,Team,Club,Friends,Other,Session,Base
-		Type string `json:"type"`
-	} `json:"group"`
-}
